feat/fix-flaky-tests: fixed more tests
